Count directory sizes as 0 if excluded
If directories are excluded or considered a mountpoint and not crossed
into, their own self stat size is still counted. This is a problem for
mountpoints which report large sizes for directories (e.g. CephFS with
recursive stats enabled), as that gets counted even if the mountpoint is
not otherwise recursed into.

A simple solution is to just not count directory sizes in the subtree
total if they are marked as excluded.